refactor; standardizing naming in issues 3
