+++ tests/tests_unit/rules/test_validation/test_formatters.py
@@ -1,6 +1,7 @@
 import pytest
 from cognite.client.data_classes.data_modeling import ContainerId, ViewId
 
+import cognite.neat.rules.issues.spreadsheet
 from cognite.neat.rules import issues as validation
 from cognite.neat.rules.issues import IssueList
 from cognite.neat.rules.issues.formatters import BasicHTML
@@ -18,13 +19,13 @@ def issues() -> IssueList:
                 input="Apple",
                 url="https://errors.pydantic.dev/2.6/v/bool_parsing",
             ),
-            validation.MultiNullableDefinitions(
+            cognite.neat.rules.issues.spreadsheet.MultiNullableError(
                 container=ContainerId("neat", "Flowable"),
                 property_name="maxFlow",
                 row_numbers={4, 5},
                 nullable_definitions={True, False},
             ),
-            validation.MissingContainerProperty(
+            validation.MissingContainerPropertyError(
                 container=ContainerId("neat", "Flowable"), property="minFlow", referred_by=ViewId("neat", "Pump", "1")
             ),
         ],
